refactor 'token/keycloak/keys' pkg to 'token/jwt'
