+++ controller/token.go
@@ -18,7 +18,7 @@ import (
 	"github.com/fabric8-services/fabric8-auth/login"
 	"github.com/fabric8-services/fabric8-auth/rest"
 	"github.com/fabric8-services/fabric8-auth/token"
-	keycloakkeys "github.com/fabric8-services/fabric8-auth/token/keycloak/keys"
+	authjwt "github.com/fabric8-services/fabric8-auth/token/jwt"
 	"github.com/fabric8-services/fabric8-auth/token/link"
 	"github.com/fabric8-services/fabric8-auth/token/provider"
 
@@ -56,7 +56,7 @@ func NewTokenController(service *goa.Service, db application.DB, auth login.Keyc
 
 // Keys returns public keys which should be used to verify tokens
 func (c *TokenController) Keys(ctx *app.KeysTokenContext) error {
-	var publicKeys keycloakkeys.JSONKeys
+	var publicKeys authjwt.JSONKeys
 	if ctx.Format != nil && *ctx.Format == "pem" {
 		publicKeys = c.TokenManager.PemKeys()
 	} else {
